agentic_ai/pyproject.toml
[tool.poetry]
name = "Stock Analysis Tool"
version = "0.1.0"
description = "Agentic AI Stock Analysis Tool"
package-mode = false

[tool.poetry.group.dev.dependencies]
python = ">=3.9, <=3.13"
crewai = {extras = ["tools"], version = "^0.85.0", python = ">=3.10,<=3.13"}
python-dotenv = {version = "^1.0.0", python = ">=3.8"}
html2text = {version = "^2024.2.26", python = ">=3.8"}
sec-api = "^1.0.20"
unstructured = {version = "<=0.10.25", python = ">=3.7"}
pyown = {version = "0.3.2", python = ">=3.11"}
langchain = {version = " 0.3.18", python = ">=3.9,<3.12"}
openai = {version = "1.63.0", python=">=3.8"}
langchain_openai = {version = " >=0.2.1,<0.3.0", python = ">=3.9.0,<4.0"}    
rapidfuzz={version="3.12.1",python=">=3.9"}
spacy = {version="3.8.4", python= ">=3.9,<3.13"}


[tool.poetry.dependencies]
fuzzywuzzy = "^0.18.0"
streamlit = {version = "^1.42.2", python = ">=3.9,<3.9.7 || >3.9.7"}
python-levenshtein = {version = "^0.26.1", python = ">=3.9"}
[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true 
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = "poetry.core.masonry.api"
